Add support for "reread" option to keybinds
If "-r" or "--reread" is specified, Vit will reload the
Taskwarrior data and refresh the screen after the command is run.
For example, if a command deletes a task, if this option is not
specified, Vit does not know to reload the data so the task will
still appear on your screen. The correct way to do this is:

  map --reread t=task % delete