Reimplement keybind framework
This commit simplifies the implementation and the interface of
the keybind framework. Note that the previous framework was
not included in a release.

The heart of the new implementation is that ":! CMD" now
executes "CMD" in the shell. This behavior was implemented in
the previous commit. This commit implements options for running
shell commands. The current specification is

  :![OPTIONS] <command>

where current options are 'r' for "reread" (this was specified
as "-r" or "--reread" in the previous framework) and 'w' for
"wait" (previously "-w" or "--wait").

Note that the space before <command> is always required so that

  :!<command>

is not valid.

There is no longer a distinction between "internal" and "external"
keybinds. Every keybind is now "internal". The reason for this is
that because of the new implementation of ":!" external commands
can be run internally. Finally, the placeholder '%' for the
currently selected task ID is now "%TASKID". The reason for this
change is that more placeholders will be implemented and instead
of searching for new symbols, they will all take the form
"%PLACEHOLDER".

As an example of changes, what was previously specified as

  map -w n=tasknote %

in .vitrc can now be specified as

  map n=:!w tasknote %TASKID\n

The 'w' is just for showing the change and would not actually be
used in this case.

Note that we remove the use of the module "Getopt::Long" because
it was only being used by the previous keybind framework.