Implement cursor movement in prompts (#1403)
Left and right cursor movements are added to prompts. There are
two parts of the code that handle prompts: searches, '/' and '?';
and other commands, /[Aafmp:]/.

The cursor must be updated on every modification of the prompt
string. This currently includes left and right movement,
backspace, and \cu.

When we want to draw the prompt and update the cursor to a point
that is not the end of the prompt string, we use the new function
'draw_prompt_cur()'. The cursor is a global variable so it is not
passed in the function.