Workaround for freezing behavior when task waits
task sometimes waits for user input that is not passed by VIT.
The following is an example:

    This is a recurring task. Do you want to modify all
    pending recurrences of this same task? (yes/no)

There are a few ways to proceed. This commit implements the first,
but the others are noted here for logging purposes.

(1) Pipe "yes" to task. It will usually be ignored, but if
    task wants STDIN it will get a handful of "yes" no matter
    what the question is.

(2) Run task in a pipe using open2() or better yet open3(). This gives
    us the file handles to STDIN.

(3) Always "wait" (this would allow the user to provide STDIN directly
    to task, but has the disadvantage that on every action the user would
    have to press "return" to go back to VIT.

Thanks to Paul for the advice.