Fix a prompt bug that prevented editing
When redefining a filter, the old string was
treated like the new prompt (so could not be edited).

The reason is that VIT passes a string to the prompt
that is meant to be parsed. The length of the prompt
(the part that should not be edited) was being based
on the entire string whereas it should be calculated
after parsing.

Thanks to Nemo Inis for reporting this bug.