Give useful error if long shortcut (see VT-81)
VIT relies on a trick to implement shortcuts. It uses ungetch to
preload the shortcut and then run it. It works well in the sense that
the shortcut should work exactly as if you had typed in each letter.
This will also be useful for testing. However, it appears there is a
limit to the number of characters you can ungetch before you need to
getch.

I hope to allow for long shortcuts in the next major release of VIT.