sudo not needed for installation into home (VT-97)
Users can now install VIT into their home directory without needing
sudo privileges. The only secondary effect is that if the user
tries to install into a directory where sudo privileges are needed,
instead of prompting for the password, a "Permission denied" error
is given.

Thanks to Ankur Sinha for the patch.