Allow Taskwarrior cmds to parse (VT-111, VT-114)
This commit allows Taskwarrior "add" and "modify" commands to
parse the user's input, such as allowing for adding a description
as well as a tag when adding a task with 'a', and for making more
than one modification when modifying a task with 'm'.

Annotations are kept as they are because there is nothing else
for Taskwarrior to interpret and the advantage is that the user
does not have to worry about quoting.

Thanks to David Patrick and Peter Novak for the reports.