Fix another invisible prompt issue
This one is in prompt_chr (e.g. triggered by completing
or deleting a task). The solution involves a recursive call
because there is no while loop for getch().

I also added debug output to log when the 410 character code
is received.

For more information, see the previous commit (45c965d3).