Reenable blinking after invisible prompt issue
In a prompt, the cursor normally blinks. However,
After a 410 character code it stopped blinking.
This commit reenables the blinking.

This commit adds on to the fix at 45c965d3, which
fixed the invisible text. In some sense, this fixes
the "invisible cursor".

This fix is not needed for prompt_chr() because blinking
is already enabled again anyway.