Improve detection of annotations (VT-123)
VIT uses a regex to detect whether a line is part of the task
description or an annotation (which is needed for allowing e.g. to
denotate a task). This regex broke detection with newer versions of
Taskwarrior because the default dateformat changed (dashes are now
used instead of dashes).

The fix is to allow the delimiter to be either a slash or dash. This
covers the two main date formats and preserves backwards
compatibility, but could still fail if a user has a different format
(by setting 'dateformat' in .taskrc). Improvements could be
considered to either parse "task show dateformat" or a different way
of detecting whether the line is part of the description or
annotation (e.g. parse the JSON export).

Thanks to Terran Air.