Add 'burndown' config and set to 'no' by default
This commit changes the behavior of VIT by not displaying the
burndown information (time until convergence) by default. The reason
for this change is that burndown can take a significant amount of
time and it should not slow down a user who just wants to view and
modify tasks.

To get the old behavior back, add the following line to the .vitrc
file:

  set burndown = no

A long-run goal is to use threads so that burndown can be run in the
background without slowing down anything.