Clean up terminal on a Perl error or warning
Before, VIT would exit on a Perl error but the terminal would be
confusing (e.g. input might be invisible). This commit changes
additional behavior in the sense that now VIT also exits if there is
a Perl warning. An example of a warning is if you add the following
to the .vitrc:

  map \c<F2>=:!wr ls

Before this commit, the binding would not be valid but would not be
reported to the user until they exited VIT. Now it is reported
immediately.

This fix is a minor part of VT-127.