Do not exit when invalid regex for search string
Search in VIT accepts regular expressions from users. It is a
feature to be able to search with regexes, but VIT should not exit
if the regex is invalid. Instead, if VIT detects an invalid regex it
searches for the fixed string. For example, if "abc\" is entered
(which is an invalid regex), the exact string "abc\" is searched
for. This behavior is consistent with Vim.

The functionality described above is implemented with a try/catch
block (so we now require Try::Tiny). It was not straightforward to
implement because we have an END block which would normally exit on
an error even if inside a try block (see e87c1232 for why we have an
END block). A new global variable is thus introduced to track
whether we are in a try/catch block.

Thanks to Rowan Thorpe.