Annotations now correctly escape any character
Single quotes are used for quoting the string and single quotes
within the string are escaped so that any character should be
allowed without concern that it will be interpreted by the shell.

Examples of characters that would cause problems before this commit
are double quotes, dollar signs, and exclamation points.