Do not exit if terminal resized to small height
When the termal is resized, VIT receives a 410 character and
redraws. However, redrawing in a terminal with very small height
(e.g. one row) fails because of underlying ncurses functions.

The fix simply involves not redrawing if the terminal is small. If
the terminal is resized to a larger size then redrawing naturally
resumes. The user likely won't notice the lack of redrawing because
to trigger this situation the terminal must be unusably small and
thus it was likely not intended by the user, or the terminal size is
in a transitional state.