initial fix for displaying multi-byte report data
This patch fixes some glitches in the display of tasks
if multi-byte characters are contained in the task
description. The glitches concerned (1) the positioning
of text in columns to the right of the task description
and (2) the background color around such text when
highlighting and de-highlighting the rows.

The patch works by using decode_utf8 before the length
function is invoked on the displayed report data.

As test case, German umlauts served.