fix for multi-byte prompt input
This patch fixes a glitch occurring when multi-byte input
is typed into the prompt line. In this situation, the cursor
previously moved to the right by more than just one position.
The typed text was displayed correctly though (no gaps in
between). This patch fixes the bug VT-100
(https://bug.tasktools.org/browse/VT-100)

The patch works by invoking the Curses window's getch()
multiple times for a single "character", if this "character"
indicates that it represents the first byte of a multi-byte
character. The whole multi-byte character is then read and
returned as a whole in a single iteration of the prompt's
main loop. This functionality is implemented in function
prompt_u8getch() in prompt.pl.

A note about the "Wide-Character-Aware Functions" in Curses
(http://search.cpan.org/~giraffed/Curses-1.32/Curses.pm).
I actually tried to use them as well (as alternative to the
slightly hack-ish prompt_u8getch() that I found online), but
did not manage to get them work. Concretely: Just using "getchar"
completely erased the screen already when I press the "a" for
adding a task. Using "$prompt_win->getchar" showed unexpected
other characters than the German umlauts I used for testing.

As test case, German umlauts as well as the cyrillic 'Ya' served.
Other special characters (like arrows or \t) are tested to still
work as before.