fix for multi-byte searches
This patch fixes searching for strings including multi-byte characters.
The patch includes a fix for the search prompt such that it no longer
advances the cursor by several positions for a single multi-byte char.
This fix is analogous to the "fix for multi-byte prompt input".
Moreover, the patch includes a fix for the search itself such that it
actually discovers matches containing multi-byte chars.

The first part of the patch uses the utf8-aware prompt for searches.
For the second part of the patch, the read input is now utf8-decoded at
a central place in the code. This second part made a previously added
decoding in the display code obsolete.

As test case, German umlauts served.