fix for multi-column characters in prompt
This patch fixes a glitch occurring when characters
are entered into the prompt that have a display width
of more than one column: Previously, the cursor position
remained too far to the left in this case.

The patch works by translating the character position
in the prompt string to the appropriate terminal column.
This translation is aware of multi-column characters,
by relying on the Text::CharWidth module.

As test case, the CJK character U+5047 was used.