fix for tabular display of multi-column chars
This patch fixes the tabular display of tasks whose
project name or task description contain multi-column
characters. Previously, the width of multi-column
characters was treated as 1 such that columns were
displayed too narrow.

The patch works by using function mbswidth (from the
Text::CharWidth module) instead of the built-in
function length to determine the width of a table
cell.

As test case, the CJK character U+5047 was used, both
for a project name as well as for a task description.