Do not abbreviate in underlying 'task' command
Use "project:..." instead of "proj:...".

This commit makes the use of the 'p' functionality in VIT robust to
a high value of the Taskwarrior config variable
"abbreviation.minimum". For example if the user sets that config
variable to 5, then in VIT if they use the 'p' key to add/modify a
project to the string "myProjectName", then instead of the project
name changing, the description will change to "proj:myProjectName".