Make search next/previous smoother
If n/N is executed before initial search, do nothing. Before, VIT
would move down a line (because it matched an empty string), which
was not intuitive. This new behavior is consistent with Vim. mutt is
different and instead gives the search prompt.

If n/N is executed after a "search not found", still try the search
again (e.g. filter could have changed) instead of searching with an
empty pattern (previous behavior). Consistent with Vim/mutt.