Allow tag list to be separated by commas (#5)
Tags can now be separated by spaces or commas.

The user might want to enter

    a, b, c

or

    a,b,c

and intend for the tags to be +a +b +c.

This input format may be intuitive to the user as it is what is
displayed when running

    task <ID> edit

This input format precludes inputting a comma as part of a tag.