Use CHANGES date instead of build date
This commit improves build reproducibility of VIT.
See https://reproducible-builds.org/ for why this is good.
This date call only works with GNU date and BSD date.

It does not really matter that the CHANGES file is not always
updated. It matters that it does not change for release tarballs
that are used to build packages. And developers can easily touch
CHANGES to update it for a test build.

Patch from Bernhard M. Wiedemann.