Remove prefixed "vit-" from VERSION
This changes the current output of "vit --version" from
"vit-1.3.dev (<git hash>)" to "1.3.dev (<git hash>)".
Note that pre-releases and final releases do not have "(<git hash>)"
appended.

This change is made partly for stylistic purposes, and also so that
if a user would like to store the VIT version in a shell variable,
it is now easier.