Fix #192: add disallowed reports with error message
TaskWarrior 2.5.2 includes a new 'timesheet' report that does not allow
configuration of anything but the report filter. This report is so
non-standard that it doesn't make any sense to try and display it in the
standard VIT report format.

Therefore, a new disallowed_reports list is kept, which removes these
reports from the valid list of generated reports, and throws an error
message with instructions on how to run the report as a shell command,
eg. ':!w task timesheet